fix(xai): normalize Responses root tool schemas
